Working with floats
added a lookup table for converting from int to float.

the first float used will take the longest and every float afterwards will take just a few milliseconds.

over a large program working with floats will be *much* faster.

smaller programs will be less efficient.

if this tends to be slower in the long run, the conversion math can always be reused.

[ticket: X]